+++ RELEASE_NOTES.md
@@ -4,9 +4,13 @@ Until AMMiSTer reaches 1.0.0, it is considered "alpha". More features and improv
 
 # Releases
 
+## Version 0.4.0
+
+- [All] When bulk adding games, if the plan path does not exist, it gets created
+
 ## Version 0.3.1
 
-- [All] Simple plan path validatin in "Bulk Add"
+- [All] Simple plan path validation in "Bulk Add"
 
 ## Version 0.3.0